(CU-85zt0qwa2 ots) migrate: operators to Composite (Tree) design pattern (see PR #34)
