--- addon/operators/reg.py
@@ -1,52 +1,0 @@
-# ##### BEGIN LICENSE BLOCK #####
-#
-# "BakeMaster" Blender Add-on (version 3.0.0)
-# Copyright (C) 2023 Kiril Strezikozin aka kemplerart
-#
-# This License permits you to use this software for any purpose including
-# personal, educational, and commercial; You are allowed to modify it to suit
-# your needs, and to redistribute the software or any modifications you make
-# to it, as long as you follow the terms of this License and the
-# GNU General Public License as published by the Free Software Foundation,
-# either version 3 of the License, or (at your option) any later version.
-#
-# This License grants permission to redistribute this software to
-# UNLIMITED END USER SEATS (OPEN SOURCE VARIANT) defined by the
-# acquired License type. A redistributed copy of this software
-# must follow and share similar rights of free software and usage
-# specifications determined by the GNU General Public License.
-#
-# This program is free software and is distributed in the hope that it will be
-# useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License in
-# GNU.txt file along with this program. If not,
-# see <http://www.gnu.org/licenses/>.
-#
-# ##### END LICENSE BLOCK #####
-
-from bpy.types import (
-    Operator,
-)
-
-
-class BM_OT_Remove_PreviewCollections(Operator):
-    """
-    Remove Preview Collections of custom icons initialized in Global
-    Properties.
-    """
-    bl_idname = 'bakemaster.remove_previewcollections'
-    bl_label = "Remove Preview Collections"
-    bl_options = {'INTERNAL'}
-
-    def execute(self, context):
-        import bpy.utils.previews as bpy_utils_previews
-        preview_collections = context.scene.bakemaster.preview_collections
-
-        for pcoll in preview_collections.values():
-            bpy_utils_previews.remove(pcoll)
-
-        preview_collections.clear()
-        return {'FINISHED'}
