+++ addon/operators/ui.py
@@ -39,7 +39,6 @@ from bpy.props import (
     StringProperty,
     BoolProperty,
 )
-from bpy_extras.io_utils import ImportHelper
 from ..utils import (
     get as bm_get,
     operators as bm_ots_utils,
@@ -50,31 +49,9 @@ from ..properties import (
     Container,
     BakeJob,
 )
-from ..labels import (
-    BM_URLs,
-)
-
 
 _walk_handler_invoked = False
 _walk_handler_invoke_time = 0
-
-
-class BM_OT_Help(Operator):
-    bl_idname = 'bakemaster.help'
-    bl_label = "Help"
-    bl_description = "Press to visit the according BakeMaster's online documentation page"  # noqa: E501
-    bl_options = {'INTERNAL'}
-
-    id: StringProperty(default="")
-
-    def invoke(self, context, _):
-        self.url = BM_URLs("latest").get(self.id)
-        return self.execute(context)
-
-    def execute(self, _):
-        from webbrowser import open as webbrowser_open
-        webbrowser_open(self.url)
-        return {'FINISHED'}
 
 
 class BM_OT_UIList_Walk_Handler(Operator):
@@ -883,1507 +860,3 @@ class BM_OT_UI_Prop_Relinquish(Operator):
 
     def invoke(self, context, _):
         return self.execute(context)
-
-
-class BM_OT_BakeJobs_Add(Operator):
-    bl_idname = 'bakemaster.bakejobs_add'
-    bl_label = "Add"
-    bl_description = "Add a new Bake Job"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    index: IntProperty(default=-1)
-
-    def invoke(self, context, _):
-        return self.execute(context)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-
-        # reduce flicker of undrawn containers
-        bm_ots_utils.disable_drag(bakemaster, bakemaster, bakemaster.bakejobs,
-                                  "bakejobs")
-
-        new_bakejob = bakemaster.bakejobs.add()
-        new_bakejob.index = bakemaster.bakejobs_len
-        new_bakejob.name = "Bake Job %d" % (new_bakejob.index + 1)
-        new_bakejob.type = bakemaster.prefs_default_bakejob_type
-
-        bakemaster.bakejobs_active_index = new_bakejob.index
-        bakemaster.bakejobs_len += 1
-
-        bm_ots_utils.indexes_recalc(bakemaster, "bakejobs")
-        return {'FINISHED'}
-
-
-class BM_OT_BakeJobs_Remove(Operator):
-    bl_idname = 'bakemaster.bakejobs_remove'
-    bl_label = "Remove"
-    bl_description = "Remove selected Bake Jobs from the list on the left"  # noqa: E501
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    index: IntProperty(default=-1)
-
-    def invoke(self, context, _):
-        if self.index == -1:
-            self.index = context.scene.bakemaster.bakejobs_active_index
-        return self.execute(context)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-
-        # remove drag_empties
-        bm_ots_utils.disable_drag(bakemaster, bakemaster, bakemaster.bakejobs,
-                                  "bakejobs", clear_selection=False)
-
-        has_selection, _ = bm_get.walk_data_multi_selection_data(
-            bakemaster, "bakejobs")
-
-        if has_selection:
-            mask = bakemaster.get_seq("bakejobs", "is_selected", bool)
-            to_remove = bakemaster.get_seq("bakejobs", "index", int)[mask]
-            size = to_remove.size
-        else:
-            bakejob = bm_get.bakejob(bakemaster, self.index)
-            if bakejob is None:
-                return {'CANCELLED'}
-            to_remove = [bakejob.index]
-            size = 1
-
-        if size < 1:
-            self.report({'INFO'}, "Nothing to remove")
-            return {'CANCELLED'}
-
-        for index in reversed(to_remove):
-            bakemaster.bakejobs.remove(index)
-
-        bm_ots_utils.indexes_recalc(bakemaster, "bakejobs")
-        bm_ots_utils.clear_multi_selection(None, bakemaster, "bakejobs")
-
-        bakemaster.bakejobs_len -= size
-        if not bakemaster.bakejobs_active_index < bakemaster.bakejobs_len:
-            bakemaster.bakejobs_active_index = bakemaster.bakejobs_len - 1
-        return {'FINISHED'}
-
-
-class BM_OT_BakeJobs_AddDropped(BM_OT_Generic_AddDropped):
-    """
-    Internal, not used in the UI, invoked from drop_name_Update.
-    """
-
-    bl_idname = 'bakemaster.bakejobs_adddropped'
-    bl_label = "New Bake Job.."
-    bl_description = "Add dropped objects into a new Bake Job"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    index: IntProperty(default=-1)
-
-    drop_name: StringProperty(default="")
-
-    def remove(self, _):
-        bpy_ops.bakemaster.bakejobs_remove('INVOKE_DEFAULT', index=self.index)
-
-    def add(self, bakemaster, add_names: list):
-        bpy_ops.bakemaster.bakejobs_add('INVOKE_DEFAULT', index=self.index)
-
-        new_bakejob = bakemaster.bakejobs[bakemaster.bakejobs_len - 1]
-        for name in add_names:
-            bpy_ops.bakemaster.containers_add('INVOKE_DEFAULT',
-                                              bakejob_index=new_bakejob.index,
-                                              new_name=name)
-
-    def invoke(self, context, _):
-        bakemaster = context.scene.bakemaster
-        try:
-            bakemaster.bakejobs[self.index]
-            if self.index == -1:
-                raise IndexError
-        except IndexError:
-            return self.execute(context)
-        else:
-            self.remove(bakemaster)
-            return {'CANCELLED'}
-
-
-class BM_OT_BakeJobs_Trash(Operator):
-    bl_idname = 'bakemaster.bakejobs_trash'
-    bl_label = "Trash"
-    bl_description = "Remove all Bake Jobs from the list on the left"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    def invoke(self, context, _):
-        return self.execute(context)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-
-        # remove drag_empties
-        bm_ots_utils.disable_drag(bakemaster, bakemaster, bakemaster.bakejobs,
-                                  "bakejobs")
-
-        [bakemaster.bakejobs.remove(index) for index in
-         reversed(range(bakemaster.bakejobs_len))]
-        bakemaster.bakejobs_active_index = -1
-        bakemaster.bakejobs_len = 0
-
-        bm_ots_utils.indexes_recalc(bakemaster, "bakejobs")
-        return {'FINISHED'}
-
-
-class BM_OT_BakeJob_Rename(Operator):
-    bl_idname = 'bakemaster.bakejob_rename'
-    bl_label = "Rename Bake Job"
-    bl_description = "Click to rename the Bake Job"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    index: IntProperty(default=-1)
-
-    new_name: StringProperty(
-        name="New name",
-        description="Enter new name",
-        default="",
-        options={'SKIP_SAVE'})
-
-    bakejob = None
-
-    def execute(self, _):
-        if self.bakejob is None:
-            return {'CANCELLED'}
-        self.bakejob.name = self.new_name
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        if self.index == -1:
-            self.report({'WARNING'}, "Internal error: Cannot resolve Bake Job")
-            return {'CANCELLED'}
-        self.bakejob = bm_get.bakejob(context.scene.bakemaster, self.index)
-        if self.bakejob is None:
-            self.report({'WARNING'}, "Internal error: Cannot resolve Bake Job")
-            return {'CANCELLED'}
-        self.new_name = self.bakejob.name
-
-        wm = context.window_manager
-        return wm.invoke_props_dialog(self, width=300)
-
-    def draw(self, _):
-        layout = self.layout
-        layout.use_property_split = False
-        layout.use_property_decorate = False
-        layout.prop(self, "new_name")
-
-
-class BM_OT_BakeJob_ToggleType(Operator):
-    bl_idname = 'bakemaster.bakejob_toggletype'
-    bl_label = "Bake Job Type"
-    bl_description = "Click to choose the Bake Job type"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    def type_objects_update(self, _):
-        if self.type_maps is not self.type_objects:
-            return
-        self.type_maps = not self.type_objects
-
-    def type_maps_update(self, _):
-        if self.type_objects is not self.type_maps:
-            return
-        self.type_objects = not self.type_maps
-
-    index: IntProperty(default=-1)
-
-    type_objects: BoolProperty(
-        name="Objects",
-        description="Bake Job will contain Objects, where each of them will contain Maps to bake",  # noqa: E501
-        default=True,
-        update=type_objects_update,
-        options={'SKIP_SAVE'})
-
-    type_maps: BoolProperty(
-        name="Maps",
-        description="Bake Job will contain Maps, where each of them will contain Objects the map should be baked for",  # noqa: E501
-        default=False,
-        update=type_maps_update,
-        options={'SKIP_SAVE'})
-
-    bakejob: BakeJob = None
-
-    def execute(self, _):
-        if self.bakejob is None:
-            return {'CANCELLED'}
-        if self.type_objects:
-            self.bakejob.type = 'OBJECTS'
-        else:
-            self.bakejob.type = 'MAPS'
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        if self.index == -1:
-            self.report({'WARNING'}, "Internal error: Cannot resolve Bake Job")
-            return {'CANCELLED'}
-        self.bakejob = bm_get.bakejob(context.scene.bakemaster, self.index)
-        if self.bakejob is None:
-            self.report({'WARNING'}, "Internal error: Cannot resolve Bake Job")
-            return {'CANCELLED'}
-        if self.bakejob.type == 'OBJECTS':
-            self.type_objects = True
-            self.type_maps = False
-        else:
-            self.type_objects = False
-            self.type_maps = True
-
-        wm = context.window_manager
-        return wm.invoke_props_dialog(self, width=100)
-
-    def draw(self, context):
-        layout = self.layout
-        layout.use_property_split = False
-        layout.use_property_decorate = False
-
-        icon_objects = bm_ui_utils.get_icon_id(context.scene.bakemaster,
-                                               "bakemaster_objects.png")
-
-        col = layout.column(align=True)
-        col.prop(self, "type_objects", icon_value=icon_objects)
-        col.prop(self, "type_maps", icon='RENDERLAYERS')
-
-
-class BM_OT_BakeJobs_Merge(Operator):
-    bl_idname = 'bakemaster.bakejobs_merge'
-    bl_label = "Merge selected Bake Jobs"
-    bl_description = "Merge selected Bake Jobs into the active one to contain all their items"  # noqa: E501
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-
-        # remove drag_empties
-        bm_ots_utils.disable_drag(bakemaster, bakemaster, bakemaster.bakejobs,
-                                  "bakejobs", clear_selection=False)
-
-        to_remove = []
-        new_active_index = -1
-        for bakejob in bakemaster.bakejobs:
-            if not bakejob.is_selected:
-                continue
-            if bakejob.index == bakemaster.bakejobs_active_index:
-                new_active_index = bakejob.index
-                continue
-            to_remove.append(bakejob.index)
-
-        if len(to_remove) < 2:
-            self.report({'INFO'},
-                        "Merge requires two or more selected Bake Jobs")
-            return {'CANCELLED'}
-
-        for index in reversed(to_remove):
-            bakemaster.bakejobs.remove(index)
-
-        bakemaster.bakejobs_active_index = new_active_index
-        bakemaster.bakejobs_len = len(bakemaster.bakejobs)
-
-        # turn off multi select
-        bakemaster.allow_multi_select = False
-        bakemaster.is_multi_selection_empty = True
-        bakemaster.multi_select_event = ''
-
-        bm_ots_utils.indexes_recalc(bakemaster, "bakejobs")
-
-        self.report({'WARNING'}, "Not implemented")
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        bakejob = bm_get.bakejob(context.scene.bakemaster)
-        if bakejob is None:
-            self.report({'WARNING'}, "Internal error: Cannot resolve Bake Job")
-            return {'CANCELLED'}
-
-        if not bakejob.is_selected:
-            self.report({'INFO'},
-                        "No active Bake Job (select it with Ctrl)")
-            return {'CANCELLED'}
-        return self.execute(context)
-
-
-class BM_OT_Containers_Add(Operator):
-    bl_idname = 'bakemaster.containers_add'
-    bl_label = "Add"
-    bl_description = "Add new Item to selected Bake Jobs.\nIf active Bake Job type is Objects, pressing this button will add all valid selected Objects in the scene to the list"  # noqa: E501
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    bakejob_index: IntProperty(default=-1)
-    new_name: StringProperty(default="")
-
-    def add(self, bakejob, name: str):
-        new_container = bakejob.containers.add()
-        new_container.index = bakejob.containers_len
-        new_container.bakejob_index = bakejob.index
-        new_container.name = name
-
-        bakejob.containers_active_index = new_container.index
-        bakejob.containers_len += 1
-
-    def invoke(self, context, _):
-        bakemaster = context.scene.bakemaster
-
-        if self.bakejob_index == -1:
-            self.bakejob_index = bakemaster.bakejobs_active_index
-        return self.execute(context)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-
-        bakejob = bm_get.bakejob(bakemaster, self.bakejob_index)
-        if bakejob is None:
-            return {'CANCELLED'}
-
-        # reduce flicker of undrawn containers
-        bm_ots_utils.disable_drag(bakemaster, bakejob, bakejob.containers,
-                                  "containers")
-
-        errors = 0
-
-        if self.new_name == "":
-            names = [object.name for object in context.selected_objects]
-        else:
-            names = [self.new_name]
-
-        if len(names) == 0:
-            self.report({'INFO'}, "Nothing is selected to add")
-            return {'CANCELLED'}
-
-        for name in names:
-            object, _, _, _, _, error_message = bm_get.object_ui_info(
-                bakemaster, context.scene.objects, name)
-
-            if object is None:
-                print("BakeMaster Object Add Error: %s" % error_message)
-                errors += 1
-                continue
-
-            self.add(bakejob, name)
-
-        if errors:
-            self.report({'WARNING'},
-                        f"{errors} Object(s) could not be added (see Console)")
-
-        bm_ots_utils.indexes_recalc(bakejob, "containers")
-        return {'FINISHED'}
-
-
-class BM_OT_Containers_Remove(Operator):
-    bl_idname = 'bakemaster.containers_remove'
-    bl_label = "Remove"
-    bl_description = "Remove selected Items from the list on the left"  # noqa: E501
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    bakejob_index: IntProperty(default=-1)
-    index: IntProperty(default=-1)
-
-    def invoke(self, context, _):
-        bakemaster = context.scene.bakemaster
-        if self.bakejob_index == -1:
-            self.bakejob_index = bakemaster.bakejobs_active_index
-
-        if self.index == -1:
-            container = bm_get.container(bm_get.bakejob(bakemaster,
-                                                        self.bakejob_index))
-            if container is not None:
-                self.index = container.index
-
-        return self.execute(context)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-
-        bakejob = bm_get.bakejob(bakemaster, self.bakejob_index)
-        if bakejob is None:
-            return {'CANCELLED'}
-
-        # remove drag_empties
-        bm_ots_utils.disable_drag(bakemaster, bakejob, bakejob.containers,
-                                  "containers", clear_selection=False)
-
-        has_selection, _ = bm_get.walk_data_multi_selection_data(
-            bakemaster, "containers")
-
-        if has_selection:
-            mask = bakejob.get_seq("containers", "is_selected", bool)
-            to_remove = bakejob.get_seq("containers", "index", int)[mask]
-            size = to_remove.size
-        else:
-            container = bm_get.container(bakejob, self.index)
-            if container is None:
-                return {'CANCELLED'}
-            to_remove = [container.index]
-            size = 1
-
-        if size < 1:
-            self.report({'INFO'}, "Nothing to remove")
-            return {'CANCELLED'}
-
-        for index in reversed(to_remove):
-            bakejob.containers.remove(index)
-
-        bm_ots_utils.indexes_recalc(bakejob, "containers")
-        bm_ots_utils.clear_multi_selection(None, bakemaster, "containers")
-
-        bakejob.containers_len -= size
-        if not bakejob.containers_active_index < bakejob.containers_len:
-            bakejob.containers_active_index = bakejob.containers_len - 1
-        return {'FINISHED'}
-
-
-class BM_OT_Containers_AddDropped(BM_OT_Generic_AddDropped):
-    """
-    Internal, not used in the UI, invoked from drop_name_Update.
-    """
-
-    bl_idname = 'bakemaster.containers_adddropped'
-    bl_label = "New Object..."
-    bl_description = "Add dropped Objects into the table"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    index: IntProperty(default=-1)
-
-    drop_name: StringProperty(default="")
-
-    def remove(self, bakemaster):
-        remove_ot = bpy_ops.bakemaster.containers_remove
-        remove_ot('INVOKE_DEFAULT',
-                  bakejob_index=bakemaster.bakejobs_active_index,
-                  index=self.index)
-
-    def add(self, bakemaster, add_names: list):
-        add_ot = bpy_ops.bakemaster.containers_add
-        for name in add_names:
-            add_ot('INVOKE_DEFAULT',
-                   bakejob_index=bakemaster.bakejobs_active_index,
-                   new_name=name)
-
-    def invoke(self, context, _):
-        bakemaster = context.scene.bakemaster
-
-        bakejob = bm_get.bakejob(bakemaster)
-        if bakejob is None:
-            return {'CANCELLED'}
-
-        try:
-            bakejob.containers[self.index]
-            if self.index == -1:
-                raise IndexError
-        except IndexError:
-            return self.execute(context)
-        else:
-            self.remove(bakemaster)
-            return {'CANCELLED'}
-
-
-class BM_OT_Containers_Trash(Operator):
-    bl_idname = 'bakemaster.containers_trash'
-    bl_label = "Trash"
-    bl_description = "Remove all Items from the list on the left"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    bakejob_index: IntProperty(default=-1)
-
-    def invoke(self, context, _):
-        return self.execute(context)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-
-        bakejob = bm_get.bakejob(bakemaster, self.bakejob_index)
-        if bakejob is None:
-            return {'CANCELLED'}
-
-        # remove drag_empties
-        bm_ots_utils.disable_drag(bakemaster, bakejob, bakejob.containers,
-                                  "containers")
-
-        [bakejob.containers.remove(index) for index in
-         reversed(range(bakejob.containers_len))]
-        bakejob.containers_active_index = -1
-        bakejob.containers_len = 0
-
-        bm_ots_utils.indexes_recalc(bakejob, "containers")
-        return {'FINISHED'}
-
-
-class BM_OT_Container_Rename(Operator):
-    bl_idname = 'bakemaster.container_rename'
-    bl_label = "Rename"
-    bl_description = "Click to rename the Object"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    index: IntProperty(default=-1)
-
-    new_name: StringProperty(
-        name="New name",
-        description="Enter new name",
-        default="",
-        options={'SKIP_SAVE'})
-
-    new_link: StringProperty(
-        name="Relink to",
-        description="Relink this item to a new Object",
-        default="",
-        options={'SKIP_SAVE'})
-
-    container: Container = None
-    bakejob_type = ''
-
-    def execute(self, context):
-        if self.container is None:
-            return {'CANCELLED'}
-
-        if self.bakejob_type == 'OBJECTS':
-            if self.container.is_group:
-                self.container.name = self.new_name
-                return {'FINISHED'}
-
-            object, _, _, _, _, error_message = bm_get.object_ui_info(
-                context.scene.bakemaster, context.scene.objects, self.new_link)
-
-            if object is None:
-                self.report({'ERROR'}, error_message)
-                return {'CANCELLED'}
-
-            self.container.name = self.new_link
-
-        elif self.bakejob_type == 'MAPS':
-            pass
-
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        bakejob = bm_get.bakejob(context.scene.bakemaster)
-        self.container = bm_get.container(bakejob, self.index)
-
-        if self.container is None or bakejob is None:
-            self.report({'WARNING'},
-                        "Internal error: Cannot resolve Container")
-            return {'CANCELLED'}
-
-        self.new_name = self.container.name
-        self.new_link = self.container.name
-        self.bakejob_type = bakejob.type
-
-        wm = context.window_manager
-        return wm.invoke_props_dialog(self, width=300)
-
-    def draw(self, _):
-        layout = self.layout
-        layout.use_property_split = False
-        layout.use_property_decorate = False
-
-        if self.container.is_group:
-            layout.prop(self, "new_name")
-        else:
-            layout.prop(self, "new_link")
-
-
-class BM_OT_Containers_Toggle_Expand(Operator):
-    bl_idname = "bakemaster.containers_toggle_expand"
-    bl_label = "Expand/Collapse"
-    bl_options = {'INTERNAL'}
-
-    bakejob_index: IntProperty(default=-1)
-    index: IntProperty(default=-1)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-        bakejob = bm_get.bakejob(bakemaster, self.bakejob_index)
-        container = bm_get.container(bakejob, self.index)
-        if bakejob is None or container is None:
-            return {'CANCELLED'}
-        elif not any([container.is_group, container.get_is_lowpoly()]):
-            return {'CANCELLED'}
-
-        container.is_expanded = not container.is_expanded
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        return self.execute(context)
-
-
-class BM_OT_Containers_GroupOptions(Operator):
-    bl_idname = "bakemaster.containers_groupoptions"
-    bl_label = "Group Type"
-    bl_description = "Change Group Options: Type and icon"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    bakejob_index: IntProperty(default=-1)
-    container_index: IntProperty(default=-1)
-
-    container: Container = None
-
-    def execute(self, _):
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        bakemaster = context.scene.bakemaster
-        bakejob = bm_get.bakejob(bakemaster, self.bakejob_index)
-        self.container = bm_get.container(bakejob, self.container_index)
-        if bakejob is None or self.container is None:
-            return {'CANCELLED'}
-
-        wm = context.window_manager
-        return wm.invoke_props_dialog(self, width=300)
-
-    def draw(self, context):
-        layout = self.layout
-        layout.use_property_split = False
-        layout.use_property_decorate = False
-
-        layout.prop(self.container, "group_type", expand=True)
-        if self.container.group_type == 'DECORATOR':
-            return
-
-        layout.separator(factor=1.0)
-        layout.label(text="Group Icon")
-        flow = layout.grid_flow(row_major=False, columns=0, even_columns=True,
-                                even_rows=False, align=True)
-
-        for color_tag, icon in bm_ui_utils.get_group_icon(
-                context.scene.bakemaster, self.container, all=True):
-            icon_ot = flow.operator(
-                "bakemaster.containers_group_seticon", text="",
-                icon_value=icon,
-                emboss=color_tag == self.container.group_color_tag)
-            icon_ot.bakejob_index = self.bakejob_index
-            icon_ot.container_index = self.container_index
-            icon_ot.new_color_tag = color_tag
-
-
-class BM_OT_Containers_Group_SetIcon(Operator):
-    bl_idname = "bakemaster.containers_group_seticon"
-    bl_label = "Change Group icon"
-    bl_description = "Change Group icon"
-    bl_options = {'INTERNAL'}
-
-    bakejob_index: IntProperty(default=-1)
-    container_index: IntProperty(default=-1)
-
-    new_color_tag: StringProperty(default="")
-
-    container: Container = None
-
-    def execute(self, _):
-        self.container.group_color_tag = self.new_color_tag
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        bakemaster = context.scene.bakemaster
-        bakejob = bm_get.bakejob(bakemaster, self.bakejob_index)
-        self.container = bm_get.container(bakejob, self.container_index)
-        if bakejob is None or self.container is None:
-            return {'CANCELLED'}
-
-        return self.execute(context)
-
-
-class BM_OT_Containers_Group(Operator):
-    bl_idname = "bakemaster.containers_group"
-    bl_label = "Group"
-    bl_description = "Group selected items"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    # NOTE:
-    # Added group item has no copy attrs call. Settings viz is carried out in
-    # the UI. If group is dictator, draw its settings, else draw nothing.
-    # For items in groups, if item is in dictator group, draw group's settings,
-    # else draw item's settings.
-
-    bakejob_index: IntProperty(default=-1)
-
-    group_insert_index = -1
-    last_selected_index = -1
-
-    def invoke(self, context, _):
-        self.group_insert_index = -1
-        self.last_selected_index = -1
-        return self.execute(context)
-
-    def eval_group(self, bakemaster, container, s_group_level,
-                   curr_group_level, p_continue_selection):
-        if self.group_insert_index == -1:
-            print(f"BakeMaster Internal Warning: group_insert_index is not defined at {self}")  # noqa: E501
-            return False, 'TRYAGAIN_ERROR'
-
-        if all([not container.ui_indent_level >= s_group_level,
-                container.is_selected]):
-            # don't clear selection on error
-            bakemaster.allow_drag_trans = False
-            return False, 'LEVEL_ERROR'
-
-        if container.is_selected:
-            if self.last_selected_index == -1:
-                self.last_selected_index = container.index
-            elif container.index != self.last_selected_index + 1:
-                # don't clear selection on error
-                bakemaster.allow_drag_trans = False
-                return False, 'ONEBLOCK_ERROR'
-            else:
-                self.last_selected_index = container.index
-
-        if container.ui_indent_level == s_group_level:
-            return container.is_selected, ''
-        elif all([container.ui_indent_level >= curr_group_level,
-                  container.ui_indent_level >= s_group_level]):
-            return p_continue_selection, ''
-
-        return False, ''
-
-    def add_group_item(self, bakejob, group_level):
-        name = "Group"
-
-        new_group = bakejob.containers.add()
-        new_group.name = name
-        # new_group.bakejob_index = bakejob.index -> in indexes_recalc()
-        new_group.is_group = True
-        new_group.ui_indent_level = group_level
-
-        bakejob.containers_len += 1
-        bakejob.containers_active_index = self.group_insert_index
-
-        bakejob.containers.move(bakejob.containers_len,
-                                self.group_insert_index)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-
-        bakejob = bm_get.bakejob(bakemaster, self.bakejob_index)
-        if bakejob is None:
-            return {'CANCELLED'}
-
-        has_selection, _ = bm_get.walk_data_multi_selection_data(
-            bakemaster, "containers")
-
-        if not has_selection:
-            self.report({'INFO'}, "Select items with Shift, Ctrl to group")
-            return {'CANCELLED'}
-
-        errors = {
-            'LEVEL_ERROR': "Cannot group with items from levels above first selected",  # noqa: E501
-            'ONEBLOCK_ERROR': "One-block selection is required to group",  # noqa: E501
-            'TRYAGAIN_ERROR': "Try again. Internal error (see console)"
-            }
-
-        to_group = []
-
-        s_group_level = -1  # inital grouping level
-        # p_group_index = -1  # index of item's parent group item
-        curr_group_level = 0  # level of item's parent group item
-        p_continue_selection = False  # continue grouping child items
-
-        for container in bakejob.containers:
-            if container.has_drop_prompt:
-                continue
-
-            if container.is_selected and s_group_level == -1:
-                s_group_level = container.ui_indent_level
-                curr_group_level = container.ui_indent_level
-                self.group_insert_index = container.index
-            elif any([s_group_level == -1, self.group_insert_index == -1]):
-                continue
-
-            if container.is_group:
-                # p_group_index = container.index
-                curr_group_level = container.ui_indent_level
-                if curr_group_level == s_group_level:
-                    p_continue_selection = container.is_selected
-            elif container.ui_indent_level <= curr_group_level:
-                # p_group_index = container.parent_group_index
-                curr_group_level = container.ui_indent_level
-                if curr_group_level == s_group_level:
-                    p_continue_selection = False
-
-            allow_group, error_id = self.eval_group(
-                bakemaster, container, s_group_level, curr_group_level,
-                p_continue_selection)
-            if error_id:
-                self.report({'INFO'}, errors[error_id])
-                return {'CANCELLED'}
-            elif allow_group:
-                to_group.append(container)
-
-        for container in to_group:
-            container.ui_indent_level += 1
-        self.add_group_item(bakejob, s_group_level)
-
-        bm_ots_utils.indexes_recalc(bakejob, "containers", parent_props=[
-            ["bakejob_index", bakejob.index]])
-        return {'FINISHED'}
-
-
-class BM_OT_Containers_Ungroup(Operator):
-    bl_idname = "bakemaster.containers_ungroup"
-    bl_label = "Ungroup"
-    bl_description = "Ungroup selected items"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    bakejob_index: IntProperty(default=-1)
-    container_index: IntProperty(default=-1)
-
-    s_ungroup_level = -1  # inital ungrouping level
-    new_active_index = -1
-
-    def ungroup(self, bakejob, container):
-        if self.s_ungroup_level == -1:
-            print(f"BakeMaster Internal Error: self.s_ungroup_level isn't assigned at {self}")  # noqa: E501
-        container.ui_indent_level = self.s_ungroup_level
-
-        if bakejob.containers_active_index != self.new_active_index:
-            bakejob.containers_active_index = self.new_active_index
-
-        if container.parent_group_index == -1:
-            return
-
-        # if parent group is dictator, copy settings from it,
-        # else do not copy anything.
-        parent_container = bakejob.containers[container.parent_group_index]
-        if parent_container.group_type == 'DECORATOR':
-            return
-
-        exclude_copy = {
-            "name": True,
-            "index": True,
-            "bakejob_index": True,
-            "is_group": True,
-            "is_expanded": True,
-            "group_type": True,
-            "group_is_texset": True,
-            "group_color_tag": True,
-            "lowpoly_index": True,
-            "lowpoly_name": True,
-            "is_cage": True,
-            "is_decal": True
-        }
-        _ = bm_props_utils_copy(parent_container, bakejob.containers,
-                                container.index, exclude_copy)
-
-    def ungroup_has_selection_errors(self, has_selection, container,
-                                     last_selected_index):
-        if not all([has_selection and container.is_selected]):
-            return last_selected_index, ''
-
-        if all([last_selected_index != -1,
-                container.index != last_selected_index + 1]):
-            return last_selected_index, 'ONEBLOCK_ERROR'
-        elif container.ui_indent_level < self.s_ungroup_level:
-            return last_selected_index, 'LEVEL_ERROR'
-        return container.index, ''
-
-    def ungroup_is_finished(self, has_selection, container, f_group_index):
-        selection_passed = all(
-            [has_selection, not container.is_selected,
-             container.ui_indent_level <= self.s_ungroup_level])
-        active_container_passed = all(
-            [not has_selection,
-             container.ui_indent_level <= self.s_ungroup_level,
-             any([
-                 all([container.is_group,
-                      container.index != f_group_index]),
-                 all([not container.is_group,
-                      container.parent_group_index != f_group_index])
-                 ])
-             ])
-        return selection_passed or active_container_passed
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-
-        bakejob = bm_get.bakejob(bakemaster, self.bakejob_index)
-        active_container = bm_get.container(bakejob, self.container_index)
-        if bakejob is None or active_container is None:
-            return {'CANCELLED'}
-
-        has_selection, _ = bm_get.walk_data_multi_selection_data(
-            bakemaster, "containers")
-
-        if not has_selection and not active_container.is_group:
-            self.report({'INFO'}, "Select at least one group to ungroup")
-            return {'CANCELLED'}
-
-        errors = {
-            'LEVEL_ERROR': "Cannot ungroup items from levels above first selected",  # noqa: E501
-            'ONEBLOCK_ERROR': "One-block selection is required to ungroup"  # noqa: E501
-        }
-
-        # when not has_selection, index of first group on self.s_ungroup_level:
-        f_group_index = -1
-        last_selected_index = -1
-
-        to_ungroup = []
-        to_remove = []
-
-        for container in bakejob.containers:
-            if container.has_drop_prompt:
-                continue
-
-            if all([self.s_ungroup_level == -1,
-                    has_selection, container.is_selected]):
-                self.s_ungroup_level = container.ui_indent_level
-                self.new_active_index = container.index
-            elif all([self.s_ungroup_level == -1, not has_selection,
-                      container.index == self.container_index]):
-                self.s_ungroup_level = container.ui_indent_level
-                self.new_active_index = container.index
-                f_group_index = container.index
-            elif self.s_ungroup_level == -1:
-                continue
-
-            # check oselection
-            last_selected_index, error_id = self.ungroup_has_selection_errors(
-                has_selection, container, last_selected_index)
-            if error_id != '':
-                self.report({'INFO'}, errors[error_id])
-                return {'CANCELLED'}
-
-            # check no more to ungroup
-            if self.ungroup_is_finished(has_selection, container,
-                                        f_group_index):
-                break
-
-            if container.is_group:
-                to_remove.append(container.index)
-                continue
-            to_ungroup.append(container)
-
-        for container in to_ungroup:
-            self.ungroup(bakejob, container)
-
-        for index in reversed(to_remove):
-            bakejob.containers.remove(index)
-        bakejob.containers_len -= len(to_remove)
-
-        bm_ots_utils.indexes_recalc(bakejob, "containers")
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        self.s_ungroup_level = -1
-        self.new_active_index = -1
-        return self.execute(context)
-
-
-class BM_OT_Subcontainers_Trash(Operator):
-    bl_idname = 'bakemaster.subcontainers_trash'
-    bl_label = "Trash"
-    bl_description = "Remove all Items from the list on the left"
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    bakejob_index: IntProperty(default=-1)
-    container_index: IntProperty(default=-1)
-
-    def invoke(self, context, _):
-        return self.execute(context)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-
-        bakejob = bm_get.bakejob(bakemaster, self.bakejob_index)
-        container = bm_get.container(bakejob, self.container_index)
-        if bakejob is None or container is None:
-            return {'CANCELLED'}
-
-        # remove drag_empties
-        bm_ots_utils.disable_drag(bakemaster, container,
-                                  container.subcontainers, "subcontainers")
-
-        [container.subcontainers.remove(index) for index in
-         reversed(range(container.subcontainers_len))]
-        container.subcontainers_active_index = -1
-        container.subcontainers_len = 0
-
-        bm_ots_utils.indexes_recalc(container, "subcontainers")
-        return {'FINISHED'}
-
-
-class BM_OT_Subcontainers_Toggle_Expand(Operator):
-    bl_idname = "bakemaster.subcontainers_toggle_expand"
-    bl_label = "Expand/Collapse"
-    bl_options = {'INTERNAL'}
-
-    bakejob_index: IntProperty(default=-1)
-    container_index: IntProperty(default=-1)
-    index: IntProperty(default=-1)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-        bakejob = bm_get.bakejob(bakemaster, self.bakejob_index)
-        container = bm_get.container(bakejob, self.container_index)
-        subcontainer = bm_get.subcontainer(bakejob, self.index)
-
-        if bakejob is None or container is None or subcontainer is None:
-            return {'CANCELLED'}
-        elif not subcontainer.is_group:
-            return {'CANCELLED'}
-
-        container.is_expanded = not container.is_expanded
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        return self.execute(context)
-
-
-class BM_OT_FileChooseDialog(Operator, ImportHelper):
-    bl_idname = 'bakemaster.filechoosedialog'
-    bl_label = "Choose a filepath"
-    bl_description = "Open a file browser, hold Shift to open the file, Alt to browse containing directory"  # noqa: E501
-    bl_options = {'INTERNAL'}
-
-    prop_name: StringProperty(default="", options={'HIDDEN'})
-
-    config_lookup: BoolProperty(default=False, options={'HIDDEN'})
-    config_action: StringProperty(default="", options={'HIDDEN'})
-
-    message: StringProperty(default="", options={'HIDDEN'})
-
-    def process_exit(self):
-        if self.message != "":
-            self.report({'INFO'}, self.message)
-        return {'FINISHED'}
-
-    def execute(self, context):
-        if os_path.isfile(self.filepath):
-            self.filepath = os_path.dirname(self.filepath)
-        # no safe check if property is invalid
-        setattr(context.scene.bakemaster, self.prop_name, self.filepath)
-
-        if not self.config_lookup:
-            return self.process_exit()
-
-        bpy_ops.bakemaster.config('EXEC_DEFAULT', action=self.config_action)
-        return self.process_exit()
-
-
-class BM_OT_Setup(Operator):
-    bl_idname = 'bakemaster.setup'
-    bl_label = "Configure the Setup"
-    bl_description = "Choose a filepath for presets, manage the config (load/save all settings into a file)"  # noqa: E501
-    bl_options = {'INTERNAL'}
-
-    def config_instruction_update(self, _):
-        default = "Config: save/load all settings & setup:"
-        self.config_instruction = default
-
-    config_instruction: StringProperty(
-        name="What is Config?",
-        description="Use config file as a super preset to save all settings, setup, presets, and tables items (e.g Bake Jobs, Objects, Maps...). BakeMaster Config file can be loaded in another .blend file, and it will still remember where to pick Objects from to bake",  # noqa: E501
-        default="Config: save/load all settings & setup:",
-        update=config_instruction_update)
-
-    def execute(self, _):
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        wm = context.window_manager
-        return wm.invoke_props_dialog(self, width=300)
-
-    def draw(self, context):
-        bakemaster = context.scene.bakemaster
-        layout = self.layout
-
-        layout.use_property_split = False
-        layout.use_property_decorate = False
-
-        if not bakemaster.config_is_attached:
-            row = layout.row(align=True)
-            row.prop(bakemaster, 'presets_filepath', text="Presets Dir")
-            p_fc = row.operator('bakemaster.filechoosedialog', text="",
-                                icon='FILEBROWSER')
-            p_fc.filepath = bakemaster.presets_filepath
-            p_fc.prop_name = "presets_filepath"
-            p_fc.message = "Filepath for Presets set successfully"
-
-            layout.separator(factor=1.0)
-
-        col = layout.column(align=True)
-        row = col.row()
-        row.prop(self, "config_instruction", text="")
-        row.enabled = False
-        c_load_text = "Load"
-
-        if bakemaster.config_is_attached:
-            cf_row = col.row(align=True)
-            cf_row.prop(bakemaster, 'config_filepath')
-            c_fc = cf_row.operator('bakemaster.filechoosedialog', text="",
-                                   icon='FILEBROWSER')
-            c_fc.filepath = bakemaster.config_filepath
-            c_fc.prop_name = "config_filepath"
-            c_fc.message = "Filepath for Config set successfully"
-            cf_row.enabled = False
-
-            c_load_text = "Reload"
-
-        row = col.row(align=True)
-
-        if bakemaster.config_is_attached:
-            row.operator('bakemaster.config', text="Save").action = 'SAVE'
-        else:
-            c_save = row.operator('bakemaster.filechoosedialog', text="Save")
-            c_save.filepath = bakemaster.config_filepath
-            c_save.prop_name = "config_filepath"
-            c_save.config_lookup = True
-            c_save.config_action = 'SAVE'
-            c_save.message = "Config saved successfully"
-
-        c_load = row.operator('bakemaster.filechoosedialog', text=c_load_text)
-        c_load.filepath = bakemaster.config_filepath
-        c_load.prop_name = "config_filepath"
-        c_load.config_lookup = True
-        c_load.config_action = 'LOAD'
-        c_load.message = "Config loaded successfully"
-
-        if bakemaster.config_is_attached:
-            row.operator('bakemaster.config', text="",
-                         icon='UNLINKED').action = 'DETACH'
-
-
-class BM_OT_Config(Operator):
-    bl_idname = 'bakemaster.config'
-    bl_label = "Config"
-    bl_description = "Load/Save/Detach Bake Configuration data. You can use config as a super preset that holds all settings and tables' items of the current BakeMaster session"  # noqa: E501
-    bl_options = {'INTERNAL', 'UNDO'}
-
-    action: EnumProperty(
-        name="Action",
-        description="Choose an action for the config. Hover over values to see descriptions",  # noqa: E501
-        default='LOAD',
-        items=[('LOAD', "Load", ""),
-               ('SAVE', "Save", ""),
-               ('DETACH', "Detach", "")],
-        options={'SKIP_SAVE'})
-
-    filepath: StringProperty(
-        name="Filepath",
-        description="Where to Save/Load a config from",
-        default="",
-        subtype='DIR_PATH')
-
-    def execute(self, context):
-        scene = context.scene
-        bakemaster = scene.bakemaster
-
-        bakemaster.config_is_attached = self.action != 'DETACH'
-        if self.filepath != "":
-            bakemaster.config_filepath = self.filepath
-        print(self.action)
-        print(self.filepath)
-
-        self.report({'WARNING'}, "Not implemented")
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        return self.execute(context)
-
-
-class BM_OT_Bake_One(Operator):
-    bl_idname = 'bakemaster.bake_one'
-    bl_label = "Bake One"
-    bl_description = "Choose to bake the active selected Map, Object, or Bake Job"  # noqa: E501
-    bl_options = {'INTERNAL'}
-
-    action: EnumProperty(
-        name="Choose an operation",
-        description="Choose an operation",
-        default='BAKEJOB',
-        items=[('MAP', "Map", "Bake the current active map"),
-               ('OBJECT', "Object", "Bake the current active object"),
-               ('BAKEJOB', "Bake Job", "Bake the current active bake job")])
-
-    def bake_poll(self, context):
-        bakemaster = context.scene.bakemaster
-        bake_is_running = bakemaster.bake_is_running
-        poll_success, message = bm_ots_utils.ui_bake_poll(bakemaster,
-                                                          bake_is_running)
-        if not poll_success:
-            self.report({'ERROR'}, message)
-            return False
-        return True
-
-    def props_set_explicit(self, bakemaster):
-        bakemaster.bake_trigger_stop = False
-        bakemaster.bake_trigger_cancel = False
-        bakemaster.bake_hold_on_pause = False
-        bakemaster.bake_is_running = True
-
-    def invoke(self, context, _):
-        if not self.bake_poll(context):
-            return {'CANCELLED'}
-
-        wm = context.window_manager
-        return wm.invoke_props_dialog(self, width=300)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-        self.props_set_explicit(bakemaster)
-        bm_ots_utils.bakehistory_add(bakemaster)
-        self.report({'WARNING'}, "Not implemented")
-        return {'FINISHED'}
-
-    def draw(self, _):
-        row = self.layout.row(align=True)
-        row.prop(self, 'action', expand=True)
-
-
-class BM_OT_Bake_All(Operator):
-    bl_idname = 'bakemaster.bake_all'
-    bl_label = "Bake All"
-    bl_description = "Bake the whole setup (all bake jobs, all maps for all objects)"  # noqa: E501
-    bl_options = {'INTERNAL'}
-
-    def bake_poll(self, context):
-        bakemaster = context.scene.bakemaster
-        bake_is_running = bakemaster.bake_is_running
-        poll_success, message = bm_ots_utils.ui_bake_poll(bakemaster,
-                                                          bake_is_running)
-        if not poll_success:
-            self.report({'ERROR'}, message)
-            return False
-        return True
-
-    def props_set_explicit(self, bakemaster):
-        bakemaster.bake_trigger_stop = False
-        bakemaster.bake_trigger_cancel = False
-        bakemaster.bake_hold_on_pause = False
-        bakemaster.bake_is_running = True
-
-    def invoke(self, context, _):
-        if not self.bake_poll(context):
-            return {'CANCELLED'}
-
-        return self.execute(context)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-        self.props_set_explicit(bakemaster)
-        bm_ots_utils.bakehistory_add(bakemaster)
-        self.report({'WARNING'}, "Not implemented")
-        return {'FINISHED'}
-
-
-class BM_OT_Bake_Pause(Operator):
-    bl_idname = 'bakemaster.bake_pause'
-    bl_label = "Pause/Resume"
-    bl_description = "Pause/Resume the bake"
-    bl_options = {'INTERNAL'}
-
-    def pause_poll(self, context):
-        bakemaster = context.scene.bakemaster
-        bake_is_running = bakemaster.bake_is_running
-        poll_success, _ = bm_ots_utils.ui_bake_poll(bakemaster,
-                                                    not bake_is_running)
-        return poll_success
-
-    def props_set_explicit(self, bakemaster):
-        is_paused = bakemaster.bake_hold_on_pause
-        bakemaster.bake_hold_on_pause = not is_paused
-
-    def invoke(self, context, _):
-        if not self.pause_poll(context):
-            return {'CANCELLED'}
-
-        return self.execute(context)
-
-    def execute(self, context):
-        self.props_set_explicit(context.scene.bakemaster)
-        self.report({'WARNING'}, "Not implemented")
-        return {'FINISHED'}
-
-
-class BM_OT_Bake_Stop(Operator):
-    bl_idname = 'bakemaster.bake_stop'
-    bl_label = "Stop"
-    bl_description = "Stop and save what's been baked"
-    bl_options = {'INTERNAL'}
-
-    def stop_poll(self, context):
-        bakemaster = context.scene.bakemaster
-        bake_is_running = bakemaster.bake_is_running
-        poll_success, _ = bm_ots_utils.ui_bake_poll(bakemaster,
-                                                    not bake_is_running)
-        return poll_success
-
-    def props_set_explicit(self, bakemaster):
-        bakemaster.bake_trigger_stop = True
-        bakemaster.bake_trigger_cancel = False
-        bakemaster.bake_hold_on_pause = False
-        bakemaster.bake_is_running = False
-
-    def invoke(self, context, _):
-        if not self.stop_poll(context):
-            return {'CANCELLED'}
-
-        return self.execute(context)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-        self.props_set_explicit(bakemaster)
-        bm_ots_utils.bakehistory_unreserve(bakemaster)
-        bakemaster.bake_trigger_stop = False
-        self.report({'WARNING'}, "Not implemented")
-        return {'FINISHED'}
-
-
-class BM_OT_Bake_Cancel(Operator):
-    bl_idname = 'bakemaster.bake_cancel'
-    bl_label = "Cancel"
-    bl_description = "Cancel - stop and erase what's been baked"
-    bl_options = {'INTERNAL'}
-
-    def cancel_poll(self, context):
-        bakemaster = context.scene.bakemaster
-        bake_is_running = bakemaster.bake_is_running
-        poll_success, _ = bm_ots_utils.ui_bake_poll(bakemaster,
-                                                    not bake_is_running)
-        return poll_success
-
-    def props_set_explicit(self, bakemaster):
-        bakemaster.bake_trigger_stop = False
-        bakemaster.bake_trigger_cancel = True
-        bakemaster.bake_hold_on_pause = False
-        bakemaster.bake_is_running = False
-
-    def invoke(self, context, _):
-        if not self.cancel_poll(context):
-            return {'CANCELLED'}
-
-        return self.execute(context)
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-        self.props_set_explicit(bakemaster)
-        bm_ots_utils.bakehistory_unreserve(bakemaster)
-        bakemaster.bake_trigger_cancel = False
-        self.report({'WARNING'}, "Not implemented")
-        return {'FINISHED'}
-
-
-class BM_OT_BakeHistory_Rebake(Operator):
-    bl_idname = 'bakemaster.bakehistory_rebake'
-    bl_label = "Rebake"
-    bl_description = "Rebake content of this bake in the history"
-    bl_options = {'INTERNAL'}
-
-    index: IntProperty(default=-1)
-
-    action: EnumProperty(
-        name="Action",
-        description="Choose how to rebake content of this bake in the history",  # noqa: E501
-        default='NEW_BAKE',
-        items=[('NEW_BAKE', "New Bake Output", "Rebake all content of this bake in the history without deleting its initially baked files"),  # noqa: E501
-               ('FULL_OVERWRITE', "Full Overwrite", "Rebake and overwrite all initially baked content of this bake in the history"),  # noqa: E501
-               ('SMART_OVERWRITE', "Smart Rebake", "Compare baked content of this bake in the history and the current setup and settings. Bake what's been added and rebake what's changed")])  # noqa: E501
-
-    def rebake_poll(self, context):
-        bakemaster = context.scene.bakemaster
-        poll_success, message = bm_ots_utils.ui_bakehistory_poll(self,
-                                                                 bakemaster)
-        if not poll_success:
-            self.report({'ERROR'}, message)
-            return False
-        bake_is_running = bakemaster.bake_is_running
-        poll_success, message = bm_ots_utils.ui_bake_poll(bakemaster,
-                                                          bake_is_running)
-        if not poll_success:
-            self.report({'ERROR'}, message)
-            return False
-        return True
-
-    def execute(self, _):
-        self.report({'WARNING'}, "Not implemented")
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        if not self.rebake_poll(context):
-            return {'CANCELLED'}
-
-        wm = context.window_manager
-        return wm.invoke_props_dialog(self, width=300)
-
-    def draw(self, _):
-        layout = self.layout
-        layout.use_property_split = False
-        layout.use_property_decorate = False
-        layout.prop(self, 'action')
-
-
-class BM_OT_BakeHistory_Config(Operator):
-    bl_idname = 'bakemaster.bakehistory_config'
-    bl_label = "Load Settings"
-    bl_description = "Replace all current settings and setup (e.g. bake jobs, objects, maps, all settings) with the settings and setup of this bake in the history"  # noqa: E501
-    bl_options = {'INTERNAL'}
-
-    index: IntProperty(default=-1)
-
-    def config_poll(self, context):
-        bakemaster = context.scene.bakemaster
-        poll_success, message = bm_ots_utils.ui_bakehistory_poll(self,
-                                                                 bakemaster)
-        if not poll_success:
-            self.report({'ERROR'}, message)
-            return False
-        if bakemaster.bake_is_running:
-            self.report({'ERROR'}, "Another bake is running")
-            return False
-        return True
-
-    def execute(self, _):
-        self.report({'WARNING'}, "Not implemented")
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        if not self.config_poll(context):
-            return {'CANCELLED'}
-
-        return self.execute(context)
-
-
-class BM_OT_BakeHistory_Remove(Operator):
-    bl_idname = 'bakemaster.bakehistory_remove'
-    bl_label = "Remove"
-    bl_description = "Remove this bake from the history (choose whether to leave its baked content)"  # noqa: E501
-    bl_options = {'INTERNAL'}
-
-    index: IntProperty(default=-1)
-
-    use_delete: BoolProperty(
-        name="Delete baked content",
-        description="Delete all baked content files of this bake (cannot be undone). If unchecked, just remove from the history",  # noqa: E501
-        default=False)
-
-    use_delete_blender_only: BoolProperty(
-        name="Blender only",
-        description="Leave baked files on the disk and only delete them from this .blend file",
-        default=False)
-
-    def remove_poll(self, context):
-        bakemaster = context.scene.bakemaster
-        poll_success, message = bm_ots_utils.ui_bakehistory_poll(self,
-                                                                 bakemaster)
-        if not poll_success:
-            self.report({'ERROR'}, message)
-            return False
-        return True
-
-    def execute(self, context):
-        bakemaster = context.scene.bakemaster
-        bm_ots_utils.bakehistory_remove(bakemaster, self.index)
-        bm_ots_utils.indexes_recalc(bakemaster, "bakehistory",
-                                    childs_recursive=False)
-        self.report({'WARNING'}, "Not implemented")
-        return {'FINISHED'}
-
-    def invoke(self, context, _):
-        if not self.remove_poll(context):
-            return {'CANCELLED'}
-
-        wm = context.window_manager
-        return wm.invoke_props_dialog(self, width=300)
-
-    def draw(self, _):
-        layout = self.layout
-        layout.use_property_split = False
-        layout.use_property_decorate = False
-        layout.prop(self, "use_delete")
-        if self.use_delete:
-            layout.prop(self, "use_delete_blender_only")
